api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6

pydantic==2.9.2
pydantic-settings==2.6.1

SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.3

python-multipart==0.0.12
passlib[bcrypt]==1.7.4
python-jose==3.3.0

httpx==0.27.2

email-validator==2.2.0
bcrypt==4.0.1
